pyproject.toml
[project]
name = "dynamo-prp-share"
version = "0.1.0"
description = "Library of assets and prompts for Ai Engineering"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "svgelements>=1.9.0",
    "cairosvg>=2.7.0", 
    "pydantic>=2.0.0",
    "click>=8.0.0",
    "pytest>=7.0.0",
    "pillow>=10.0.0"
]
